requirements.txt
streamlit
pandas
google-cloud-aiplatform
chromadb
openai
pysqlite3-binary
python-dotenv
requests